refactor: OffersSection and related components
to new file
